+++ tests/test_migrations/null_field_removal/0001_initial.py
@@ -19,4 +19,20 @@ class Migration(migrations.Migration):
                 ("bar", models.IntegerField()),
             ],
         ),
+        migrations.CreateModel(
+            "Bar",
+            fields=[
+                (
+                    "id",
+                    models.AutoField(
+                        auto_created=True,
+                        primary_key=True,
+                        serialize=False,
+                        verbose_name="ID",
+                    ),
+                ),
+                ("name", models.CharField(unique=True)),
+            ],
+            options={"managed": False},
+        ),
     ]